added profile page

+++ src/main/java/com/softserve/dao/GamesDao.java
@@ -11,6 +11,6 @@ public interface GamesDao {
     boolean add(Game game);
     boolean update(Game game);
     void deleteById(long gameId);
-    double calculateAverageMark(long gameId);
+    float calculateAverageMark(long gameId);
 }
