+++ src/main/java/com/softserve/dto/GameDto.java
@@ -3,13 +3,13 @@ package com.softserve.dto;
 import com.softserve.models.Game;
 
 public class GameDto extends Game {
-    private double avgMark;
+    private float avgMark;
 
     public double getAvgMark() {
         return avgMark;
     }
 
-    public void setAvgMark(double avgMark) {
+    public void setAvgMark(float avgMark) {
         this.avgMark = avgMark;
     }
 }
